metadata-web.test.ts: move in `describe("fetchPageTitle from the web"`, from metadata.test.ts
