+++ tests/utils/metadata.test.ts
@@ -231,236 +231,6 @@ describe("fetchPageTitle", () => {
   });
 });
 
-// Real web tests - these may be unstable due to network conditions
-// To run these tests, set environment variable: RUN_WEB_TESTS=true
-// Note: microlink.io API may be rate-limited or require authentication for consistent results
-//
-// IMPORTANT: These tests may show different results than the actual plugin because:
-// 1. LogSeq plugin runs in browser environment with full fetch API support
-// 2. Test environment (Node.js/Vitest) may have limited fetch capabilities
-// 3. CORS restrictions may apply differently in test vs browser environments
-// 4. The plugin logs show API working: "Generated markdown: [A privacy-first, open-source knowledge base](https://logseq.com/)"
-//
-// Usage:
-//   npm test                    # Runs only mock tests (fast, reliable)
-//   RUN_WEB_TESTS=true npm test # Runs both mock and web tests (slower, realistic)
-describe("fetchPageTitle from the web", () => {
-  // Test data with real URLs and expected metadata
-  const realWebTests = [
-    {
-      url: "https://github.com",
-      expectedData: {
-        titlePattern: /github/i,
-      },
-    },
-    {
-      url: "https://stackoverflow.com",
-      expectedData: {
-        titlePattern: /Newest Questions/i,
-      },
-    },
-    {
-      url: "https://www.npmjs.com",
-      expectedData: {
-        titlePattern: /npm/i,
-      },
-    },
-    {
-      url: "https://vitest.dev",
-      expectedData: {
-        titlePattern: /Vitest/,
-      },
-    },
-    {
-      url: "https://logseq.com",
-      expectedData: {
-        titlePattern: /A privacy-first, open-source knowledge base/,
-      },
-    },
-  ];
-
-  // Set longer timeout for real web requests
-  const webTestTimeout = 10000; // 10 seconds
-
-  // Skip all web tests unless explicitly enabled
-  if (!process.env.RUN_WEB_TESTS) {
-    it("web tests are skipped", () => {
-      console.log("Skipping web tests (set RUN_WEB_TESTS=true to enable)");
-      if (process.env.VERBOSE_METADATA) {
-        console.log("💡 For verbose output, set VERBOSE_METADATA=true");
-      }
-    });
-    return;
-  }
-
-  const isVerbose = process.env.VERBOSE_METADATA === "true";
-
-  if (isVerbose) {
-    console.log("🔍 Running web tests in VERBOSE mode");
-    console.log("📋 Test configuration:");
-    console.log("  - RUN_WEB_TESTS:", process.env.RUN_WEB_TESTS);
-    console.log("  - VERBOSE_METADATA:", process.env.VERBOSE_METADATA);
-  }
-
-  // Restore real fetch for web tests
-  beforeEach(() => {
-    // Import undici fetch for web tests
-    const {fetch: undiciFetch} = require("undici");
-    global.fetch = undiciFetch;
-  });
-
-  afterEach(() => {
-    // Restore mock fetch
-    global.fetch = mockFetch;
-  });
-
-  realWebTests.forEach(({url, expectedData}) => {
-    it(
-      `should fetch real metadata from ${url}`,
-      async () => {
-        const isVerbose = process.env.VERBOSE_METADATA === "true";
-
-        if (isVerbose) {
-          console.log(`\n🧪 Testing URL: ${url}`);
-          console.log(`⏱️  Test started at: ${new Date().toISOString()}`);
-        }
-
-        let result;
-        try {
-          result = await fetchPageTitle(url);
-
-          if (isVerbose) {
-            console.log(`✅ Test completed successfully`);
-            console.log(`📊 Result analysis:`);
-            console.log(`   - Title: "${result.title}"`);
-            console.log(`   - Has Icon: ${result.hasIcon}`);
-            console.log(`   - Note: ${result.note || "none"}`);
-          } else {
-            console.log(`Fetched metadata for ${url}:`, result);
-          }
-        } catch (error) {
-          if (isVerbose) {
-            console.log(`❌ Test failed with error:`, error);
-          } else {
-            console.warn(`Web test failed for ${url}:`, error);
-          }
-          // In case of network failure, just verify we get a fallback
-          expect(result || {title: url}).toEqual({
-            title: url,
-          });
-          return;
-        }
-
-        // Verify basic structure
-        expect(result).toHaveProperty("title");
-        expect(typeof result.title).toBe("string");
-        expect(result.title.length).toBeGreaterThan(0);
-
-        // Check if we got actual metadata or fallback
-        const gotActualMetadata = result.title !== url;
-
-        if (gotActualMetadata) {
-          // If we got real metadata, verify it matches expectations
-          if (expectedData.titlePattern) {
-            expect(result.title).toMatch(expectedData.titlePattern);
-          }
-
-          console.log(`✅ Real web test passed for ${url} (got metadata):`, {
-            title: result.title,
-            hasIcon: !!result.icon,
-          });
-        } else {
-          // If we got fallback (URL as title), that's also valid
-          expect(result.title).toBe(url);
-          console.log(`⚠️ Real web test passed for ${url} (fallback to URL):`, {
-            title: result.title,
-            hasIcon: !!result.icon,
-            note: "API may be rate-limited or unavailable",
-          });
-        }
-
-        // If icon is present, verify it's a valid URL
-        if (result.icon) {
-          expect(typeof result.icon).toBe("string");
-          expect(result.icon).toMatch(/^https?:\/\//);
-        }
-      },
-      webTestTimeout
-    );
-  });
-
-  it("should handle non-existent domains gracefully", async () => {
-    const invalidUrl = "this-domain-definitely-does-not-exist-12345.com";
-    const result = await fetchPageTitle(invalidUrl);
-
-    // Should fallback to URL as title when fetch fails
-    expect(result).toEqual({
-      title: invalidUrl,
-      hasIcon: false,
-      note: "API may be rate-limited or unavailable",
-    });
-  }, 5000);
-
-  it("should handle slow responses within timeout", async () => {
-    // Test with a service that might be slow but reliable
-    const result = await fetchPageTitle("https://httpbin.org/delay/2");
-
-    // Should still get some result even if slow
-    expect(result).toHaveProperty("title");
-    expect(typeof result.title).toBe("string");
-  }, 15000);
-
-  // Manual test for when API is working - uncomment and run individually
-  it.skip("manual test with working API (uncomment to test)", async () => {
-    // This test can be enabled manually to verify the API is working
-    // Uncomment the line below and run this specific test when needed
-    // it.only("should fetch real metadata when API is responsive", async () => {
-
-    const testUrl = "https://example.com";
-    const result = await fetchPageTitle(testUrl);
-
-    console.log("Manual API test result:", result);
-
-    // Basic verification
-    expect(result).toHaveProperty("title");
-    expect(typeof result.title).toBe("string");
-
-    // When API is working, we should get more than just the URL back
-    // This assertion will fail if API is down (expected)
-    if (result.title !== testUrl) {
-      expect(result.title.length).toBeGreaterThan(testUrl.length);
-      console.log("✅ API is working and returned metadata");
-    } else {
-      console.log("⚠️ API returned fallback (URL) - may be rate limited");
-    }
-  }, 10000);
-
-  // Debug test to investigate API response
-  it.skip("debug API response (enable manually)", async () => {
-    const testUrl = "https://logseq.com";
-
-    try {
-      const response = await fetch(
-        `https://api.microlink.io/?url=${encodeURIComponent(testUrl)}`
-      );
-
-      console.log("Raw response status:", response.status);
-      console.log(
-        "Raw response headers:",
-        Object.fromEntries(response.headers.entries())
-      );
-
-      const data = await response.json();
-      console.log("Raw API response data:", JSON.stringify(data, null, 2));
-
-      const result = await fetchPageTitle(testUrl);
-      console.log("Processed result:", result);
-    } catch (error) {
-      console.error("Debug test error:", error);
-    }
-  }, 15000);
-});
-
 describe("processBlockContentForURLs", () => {
   beforeEach(() => {
     mockFetch.mockClear();
